ativ7

+++ atividade6.py
@@ -1,7 +1,7 @@
 soma = 0
 
 while True:
-	numero = float(input(f"Digite seu número ou 0 para sair: "))
+	numero = int(input(f"Digite seu número ou 0 para sair: "))
 	if numero == 0:
 		break
 	soma += numero
